Make fixes for workout and change sendNotfication argumetns

+++ controllers/user.js
@@ -4,6 +4,8 @@ const User = require("../models/User");
 const Workout = require("../models/Workout");
 const Tournament = require("../models/Tournament");
 const increaseRP = require("../utils/increaseRP");
+const Token = require("../models/Token");
+const sendNotification = require("../utils/sendNotification");
 
 /** Get user by id
  * @url GET /user/:id
@@ -198,12 +200,52 @@ const deleteUser = async (req, res) => {
 
         for (let i = 0; i < user.workouts.length; i++) {
             const workout = await Workout.findById(user.workouts[i]);
+
+            // User is an owner
             if (workout.user.toString() === userID) {
-                await workout.deleteOne();
-            } else {
+                workout.user = null;
+
+                // Company is empty
+                if(workout.company.length === 0)
+                {
+                    await workout.deleteOne();
+                }
+                // Company is not empty
+                else
+                {      
+                    // Workout has not ended yet
+                    if(!workout.endTime)
+                    {
+                        workout.user = workout.company[0];
+
+                        // Deletes first element in array (new owner)
+                        workout.company.shift();
+
+                        const newOner = await User.findById(workout.user);
+                        await sendNotification([newOner._id],`You have become the owner of "${workout.title}" workout!`);
+                    }
+                 
+                    // If workout has ended we do nothing
+
+                    await workout.save();
+                }
+            }
+            // User in not an owner 
+            else {
+
                 workout.company = workout.company.filter((company) => company.toString() !== userID);
                 workout.access = workout.access.filter((access) => access.toString() !== userID);
-                await workout.save();
+
+                // There is no owner and user is the only company
+                if(!workout.user && workout.company.length===0)
+                {
+                    await workout.deleteOne();
+                }
+                else
+                {
+                    await workout.save();
+                }
+                        
             }
         }
 
@@ -219,6 +261,11 @@ const deleteUser = async (req, res) => {
                 await tournament.save();
             }
         }
+
+        await Token.deleteMany({ user: userID });  
+        res.cookie("jwt", "", {
+            maxAge: 1,
+          });
 
         await user.deleteOne();
         res.status(StatusCodes.OK).json({ message: "User deleted successfully" });
